telegram-bot/requirements.txt
google-generativeai>=0.8.0
python-telegram-bot>=21.0
